refactor(api): move the provider handlers out of admin.go
admin.go held the Handler wiring, route registration and auth middleware
alongside the provider CRUD handlers. The provider handlers
(CreateProvider, ListProviders, GetProvider, UpdateProvider,
DeleteProvider, the identity/URL-shape acceptance helpers,
providerTypeAllowsEmptyKey and isForeignKeyViolation) now live in
providers.go, and their 46 tests move from admin_test.go to
providers_test.go. Same package, pure moves, imports pruned; admin.go
goes from 1096 to 596 lines and admin_test.go from 2332 to 1025, so
both entries leave the size allowlist.
